requirements.txt
# Web framework & Utils
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
PyYAML>=6.0.1
httpx>=0.27.0
Pillow>=10.3.0

# Core Deep Learning & LLM tools
torch>=2.3.0
transformers>=4.37.0
diffusers<0.26.0
huggingface_hub<0.23.0
accelerate>=0.21.0
bitsandbytes>=0.41.0
qwen-vl-utils>=0.0.1

# Evaluation Models & APIs
image-reward>=1.0
hpsv2>=1.1.0
trendspyg>=0.1.0
gdown>=5.1.0
clip-anytorch>=2.6.0
openai>=1.12.0